Modificacion a endpoints de entidad Historial

+++ src/main/java/com/Vet/VetBackend/clinica/app/services/HistorialService.java
@@ -11,6 +11,6 @@ public interface HistorialService {
     CompletableFuture<HistorialDto> actualizarHistorial(Long id, HistorialDto historialDTO);
     CompletableFuture<List<HistorialDto>> obtenerTodos();
     CompletableFuture<HistorialDto> obtenerPorId(Long id);
-    CompletableFuture<List<HistorialDto>> obtenerHistorialesPorMascota(Long mascotaId);
+    CompletableFuture<List<HistorialDto>> obtenerHistorialesPorMascota(Integer mascotaId);
 }
